requirements.txt
requests==2.31.0
botbuilder-integration-aiohttp>=4.14.5
python-dotenv